sort positions

+++ src/team.c
@@ -23,7 +23,7 @@ double avgRating(team* t) {
   if (t == NULL) return 0.0;
   double sum = 0;
   int ratings_n = 0;
-  for (int pI = 0; pI < (int)t->size; pI++) {
+  for (size_t pI = 0; pI < t->size; pI++) {
     double r = rating(t->players[pI]);
     if (fabs(r) > 1e-6f) {
       sum += r;
